fix: Workaround I/O errors on some NVS drivers
At least for the ESP driver, a `nvs_read` from a NULL pointer gives I/O
error, even when `len` is set to 0. My understanding of the API docs is
that it should not be the case, but my painfully slow discovery tells
otherwise.

Workaround this by allocating 1 byte first, and then eventually
reallocating to the required length.

Changelog: None
Ticket: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>