fix: Fix logging crashing.
Weirdly logging was casing kernel crashes. I managed to trace it down
to the calls to `do_swap()`. Initially I thought that there was something
wrong with my simple scheduler implementation (more on this is other
commits) but managed to figure out that debugging itself was the root cause
of the issue.

Signed-off-by: Marcin Pasinski <marcin.pasinski@northern.tech>